ML-predictor/requirements.txt
flask==3.0.0
pandas
numpy
scikit-learn
joblib
psycopg2-binary
tensorflow==2.20.0
